tests/wpt/web-platform-tests/offscreen-canvas/shadows/2d.shadow.attributes.shadowColor.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.shadow.attributes.shadowColor.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.shadow.attributes.shadowColor.invalid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.shadowColor = '#00ff00';
ctx.shadowColor = 'bogus';
_assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");
ctx.shadowColor = '#00ff00';
ctx.shadowColor = 'red bogus';
_assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");
ctx.shadowColor = '#00ff00';
ctx.shadowColor = ctx;
_assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");
ctx.shadowColor = '#00ff00';
ctx.shadowColor = undefined;
_assertSame(ctx.shadowColor, '#00ff00', "ctx.shadowColor", "'#00ff00'");

t.done();

});
</script>
